lab3/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
passlib[bcrypt]
bcrypt
python-jose[cryptography]
python-multipart